move utils_new_frontends.py into utils.py
